// src/constants/types.ts
export enum PokemonTypes {
    normal = 'normal',
    fire = 'fire',
    water = 'water',
    electric = 'electric',
    grass = 'grass',
    ice = 'ice',
    fighting = 'fighting',
    poison = 'poison',
    ground = 'ground',
    flying = 'flying',
    psychic = 'psychic',
    bug = 'bug',
    rock = 'rock',
    ghost = 'ghost',
    dragon = 'dragon',
    dark = 'dark',
    steel = 'steel',
    fairy = 'fairy'
}